requirements.txt
torch>=2.2.0
openai-whisper
librosa
jiwer
numpy
pyyaml
